+++ com.unity.cinemachine/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Unreleased
 
 ### Bugfixes
+- Deoccluder did not always properly reset its state.
 
 ### Changed
 - Cinemachine Shot Editor longer provides UX to create cameras when editing a prefab.